Structural changes for UserRole + UserRoleProvider
